refactor: #114 - useProfileFavoritesSearch 훅 이름 변경
