+++ src/hooks/index.ts
@@ -2,5 +2,5 @@ export { default as useModal } from '@/hooks/useModal'
 export { default as useQueryString } from '@/hooks/useQueryString'
 export { default as useSortParam } from '@/hooks/useSortParam'
 export { default as useCategoryParam } from '@/hooks/useCategoryParam'
-export { default as useProfileFavoritesSearch } from '@/hooks/useProfileFavoritesSearch'
+export { default as useProfileSpacesSearch } from '@/hooks/useProfileSpacesSearch'
 export { default as useFollowUser } from '@/hooks/useFollowUser'
